requirements.txt
# Core dependencies
openai>=1.3.0
anthropic>=0.7.0
pydantic>=2.0.0
python-dotenv>=1.0.0
typer>=0.9.0
rich>=13.0.0
typing-extensions>=4.5.0

# Web Dashboard Dependencies - Phase 5A.4
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
websockets>=12.0
jinja2>=3.1.0
python-multipart>=0.0.6

# Static file serving for frontend
aiofiles>=23.0.0 